ignite registry

+++ smqtt-common/src/main/java/io/github/quickmsg/common/interate1/topic/Topics.java
@@ -52,5 +52,6 @@ public interface Topics<T> extends IntegrateGetter {
      *
      * @return counts
      */
-    Integer counts();
+    Long counts();
+
 }
